Update first-steps-with-django.rst
Setting DJANGO_SETTINGS_MODULE in celery.py doesn't work if you have multiple settings files you interact with. The instructions made it sound like there is an option you can pass to 'celery worker' to tell it the Django settings to use (similar to 'manage.py --settings=...'). This makes it clear how that information is "passed".